// src/app/models/contact.model.ts
export class Contact {
    id: number;
    nome: string;
    cognome: string;
    telefono: number;
    telefono2: number;
    email: string;
    attivita: string;
    indirizzo: string;
    azienda: string;
    figura_professionale: string;
    messaggio: string;
    letto: number;
    provincia: string;
    cap: string;
    data_contatto: Date;
    richiamare_il: Date;
    pagina_provenienza: string;
    offerta: string;
    richiama: string;
    ordinato: string;
    data_ordine: Date;
    quando: string;
    info_agg: string;
    colore: string;
    materiale: string;
    risponde: string;
    interessato: string;
    tassativo: string;
    offerta_inviata: string;
    ricontattare: string;
    vacrd: number;
    username: string;
    password: string;
    non_risponde: string;
    richiama_chi: string;
    privacy: string;
    upload_folder: string;
    ts: Date;
    landing: number;
    id_cartella: number;
    interesse: string;
    consenso_trattamento: number;
    consenso_commerciale: number;
    landing_arrivo: string;
    email_schedule_sent: number;
    invio_app_num: number;
    date_invio_app: Date;
    data_inserimento_utente: Date;
    n_cont: number;
    type: string;
    descrizione: string;
    double: Number;
    appuntamento: Date;
    data_ultimo_contatto: Date;
    id_idraulico: number;
    tipo_collaborazione: string;
    mostra_idraulico: boolean;
    startApp: Date;
    data_f:Date;
    valutazione:number;
}